Issue 15329: old thumbnails deleted before new ones are verified to work Issue 15382: Memory leak in attachment reports
Fix some resource closing inconsistencies; add close() validation and better documentation to stream-taking utility methods

CheckedInputStream (and potentially handy base class, InputStreamWrapper) that validates whether close() was called prior to finalization (like we do with ResultSetImpl)

Use ImageUtil.THUMBNAIL_HEIGHT instead of magic numbers for scaling thumbnail images

+++ api/src/org/labkey/api/attachments/DocumentConversionService.java
@@ -30,11 +30,11 @@ import java.io.OutputStream;
 // Expose some interesting conversion methods via this service.
 public interface DocumentConversionService
 {
-    // Returns null if requested page doesn't exist
+    // Closes the passed in pdfStream. Returns null if requested page doesn't exist.
     @Nullable BufferedImage pdfToImage(InputStream pdfStream, int page);
-    // Returns null if requested page doesn't exist
+    // Closes the passed in pdfStream. Returns null if requested page doesn't exist.
     @Nullable BufferedImage pdfToImage(InputStream pdfStream, int page, int bufferedImageType, int resolution);
 
     public void svgToPng(String svg, OutputStream os) throws TranscoderException;
-    public void svgToPng(String svg, OutputStream os, @Nullable Float height) throws TranscoderException;
+    public void svgToPng(String svg, OutputStream os, @Nullable Double height) throws TranscoderException;
 }
